+++ tests/api_resources/v1/test_alerts.py
@@ -44,10 +44,12 @@ class TestAlerts:
             ],
             customer_id="4db51251-61de-4bfe-b9ce-495e244f3491",
             evaluate_on_create=True,
-            group_key_filter={
-                "key": "key",
-                "value": "value",
-            },
+            group_values=[
+                {
+                    "key": "key",
+                    "value": "value",
+                }
+            ],
             invoice_types_filter=["PLAN_ARREARS, SCHEDULED, USAGE, CORRECTION, CREDIT_PURCHASE, or SEAT_PURCHASE"],
             plan_id="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
             uniqueness_key="x",
@@ -154,10 +156,12 @@ class TestAsyncAlerts:
             ],
             customer_id="4db51251-61de-4bfe-b9ce-495e244f3491",
             evaluate_on_create=True,
-            group_key_filter={
-                "key": "key",
-                "value": "value",
-            },
+            group_values=[
+                {
+                    "key": "key",
+                    "value": "value",
+                }
+            ],
             invoice_types_filter=["PLAN_ARREARS, SCHEDULED, USAGE, CORRECTION, CREDIT_PURCHASE, or SEAT_PURCHASE"],
             plan_id="182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e",
             uniqueness_key="x",