+++ tests/api_resources/v1/test_contracts.py
@@ -319,6 +319,11 @@ class TestContracts:
                             "product_tags": ["string"],
                         }
                     ],
+                    "subscription_config": {
+                        "apply_seat_increase_config": {"is_prorated": True},
+                        "subscription_id": "subscription_id",
+                        "allocation": "POOLED",
+                    },
                     "temporary_id": "temporary_id",
                 }
             ],
@@ -355,6 +360,11 @@ class TestContracts:
                             "product_tags": ["string"],
                         }
                     ],
+                    "subscription_config": {
+                        "apply_seat_increase_config": {"is_prorated": True},
+                        "subscription_id": "subscription_id",
+                        "allocation": "POOLED",
+                    },
                     "temporary_id": "temporary_id",
                 }
             ],
@@ -446,6 +456,7 @@ class TestContracts:
                     "ending_before": parse_datetime("2019-12-27T18:11:19.117Z"),
                     "name": "name",
                     "starting_at": parse_datetime("2019-12-27T18:11:19.117Z"),
+                    "temporary_id": "temporary_id",
                 }
             ],
             total_contract_value=0,
@@ -1627,6 +1638,11 @@ class TestAsyncContracts:
                             "product_tags": ["string"],
                         }
                     ],
+                    "subscription_config": {
+                        "apply_seat_increase_config": {"is_prorated": True},
+                        "subscription_id": "subscription_id",
+                        "allocation": "POOLED",
+                    },
                     "temporary_id": "temporary_id",
                 }
             ],
@@ -1663,6 +1679,11 @@ class TestAsyncContracts:
                             "product_tags": ["string"],
                         }
                     ],
+                    "subscription_config": {
+                        "apply_seat_increase_config": {"is_prorated": True},
+                        "subscription_id": "subscription_id",
+                        "allocation": "POOLED",
+                    },
                     "temporary_id": "temporary_id",
                 }
             ],
@@ -1754,6 +1775,7 @@ class TestAsyncContracts:
                     "ending_before": parse_datetime("2019-12-27T18:11:19.117Z"),
                     "name": "name",
                     "starting_at": parse_datetime("2019-12-27T18:11:19.117Z"),
+                    "temporary_id": "temporary_id",
                 }
             ],
             total_contract_value=0,